requirements.txt
# Jupyter requires  pip3 install jupyterlab "ipywidgets>=7.5"
PyMySQL==1.0.2
roundrobin==0.0.2
slackclient==2.9.3
pid==3.0.4
tabulate==0.8.9
Cython==0.29.26
ruptures==1.1.6
numpy>=1.20.3
pandas==1.2.4
requests==2.25.1
PyYAML==5.4.1
plotly==4.14.3
seaborn==0.11.1
anytree==2.8.0
dacite==1.6.0
nbstripout==0.5.0
prometheus-client==0.11.0
MarkupSafe==2.0.1
cryptography==36.0.0
asyncssh==2.10.1
jinja2>=3.1.2
prometheus-api-client==0.5.0
xmltodict==0.13.0
papermill==2.3.4
jupyterlab==3.4.4
psycopg2-binary==2.9.3
import-ipynb==0.1.3
grafana-client==3.1.0